Fix a comment that misrepresents the root marking algorithm.
During the initial root marking nothing is pushed on the mark stack.
Marked roots are pushed during the scanning phase following root
marking.  Unfortunately, the comment above the root marking callback
incorrectly claimed that it pushed objects on the mark stack.

In addition, this change renames the marking callback for the re-mark
phase to more accurately reflect its limited usage.

Change-Id: I38161fa723e00e14bd8d047a2e25cdbfd215a2c6