Do not commingle application and zygote references when sweeping.
When garbage objects are identified references to these objects are
passed to a function which, in the case of an application heap,
releases their storage to the mspace or, in the case of a zygote heap,
counts their freed memory without touching the underlying heap.

This freeing process is performed in bulk.  To attenuate the cost of
calling and to allow for more efficient merging of adjacent unused
blocks of memory, a buffer is filled with adjacent references to free.
The freeing mechanism is chosen based on the first reference in the
buffer.  If the first reference is an application heap object all
objects in the buffer have their storage released to the underlying
mspace.  If the first object is a zygote object, only the free space
counting is performed.  Assertions check if this assumption is sound.

The previous change to the sweep did not treat the various regions of
the heap specially and therefore allowed the freeing buffer to have
its contained pointers span both heaps.  This change restores some of
the functionality which kept references from either heap separate.

Change-Id: Ibc563dcf46ff1681596664f04a2a85a4d29eb5e1