Periodically reset the thin lock polling interval.
In a worst case scenario it was possible for a thread waiting to
inflate a lock to get stuck polling once a second.  With this change
the waiting thread will reset its delay back to the minimum delay
after reaching a 1 second delay.

This change also replaces a call to the obsolete usleep function with
a call to nanosleep.

Change-Id: I25bc3e3907ec070250e1d7821b7770a8d1ba9627