Hook up the preexisting RawDexFile opener.
This one is similar in functionality to what's in JarFile.

I also made a couple minor (whitespace/comment/logging) changes to
JarFile where I spotted the opportunities.

This change makes it so that, if you happen to put a plain dex file on
your classpath (e.g., "dalvikvm -classpath /data/local/blort.dex
Blort"), it will actually be able to be used. This was documented as
working, but it was a lie...until now.

Change-Id: I2c93a8fdb40afa2d920904ab1235b7909e5837f8