Factor debug locals into liveness.
If the debugger asks for a reference held in a local variable, we know
that the reference is valid because the stack is part of the root set.
With liveness analysis, it's possible for a local to be considered dead
and get GCed.  Since we don't currently have a way to deal with that
in the debugging code, we just mark debugger-visible locals as live.

Bug 2534655

Change-Id: I5b217f20a76364f1f6a89499058de6ec4eb07b7d