Simplify the interface for waiting for a concurrent GC to complete.
Previously, there was no guarantee that a GC would have completed when
the waiting routine returned.  This was an oversight, and the workaround
became to wait in a loop that tested the condition that the GC running
flag was not set.  The right fix was always to push the loop into the
waiting routine itself to keep the interface simple.  This change backs
out the hack, moving the waiting loop to where it belongs.

Change-Id: Idc6ab6e5d4379a5da27e9e115d23d4123c7cacc4