Guard against uninitialized objects when counting instances.
Objects marked in the live object bitmap are not guaranteed to have
their class object pointer installed.  With this change the subclass
check is short circuited when an NULL class pointer is encountered.
Previously, the class pointer was blindly assumed to be valid which
could cause the instance-of check to abort the VM.

Change-Id: I2d8810f1efd7643015e3ce3cf1dd93d90d330eee