[JIT] Fix for 3385583: Performance variance
Closes a window in which the "interpret-only" templace could get chained
to an existing trace while the intended translation was under construction.
Note that this CL also introduces some small, but fundamental changes in trace
formation:

1.  Previouosly, when an exception or other trace terminating event
occurred during trace formation, the entire trace was abandoned.  With this
change, we instead end the trace at the last successful instruction.

2.  We previously allowed multiple attempts (perhaps by multiple threads)
to form a trace compilation request for a dalvik PC.  This was done in an
attempt to allow recovery from compiler failures.  Now we enforce a new rule:
only the thread that wins the race to allocate an entry in the JitTable will
form the trace request.

3.  In a (probably misguided) attempt avoid unnecessary contention, we
previously allowed work order enqueue requests to be dropped if a requester
did not aquire TableLock on first attempt (assuming that if the trace were
hot, it would be requested again).  Now we block on enqueue.

Change-Id: I40ea4f1b012250219ca37d5c40c5f22cae2092f1