Tweak the 64-bit load/store code
The old version would either move data through a union, call memcpy(),
or just cast the pointer to a wider type and dereference it.  The
latter turns out to be an aliasing violation, and it recently started
causing test failures on the "host" build.

We now use memcpy() for x86 and unions for ARM.

Bug 3431820

Change-Id: I302a7f49f7ae88ac96b8f7fef3d9260ac64d631b