Ensure a dalvik-cache directory exists before running dalvik.
The dalvik tests require a dalvik-cache directory otherwise dalvikvm
will fail with an unhelpful error message.  On the device it is safe
to assume a dalvik-cache directory exists but this is not a safe
assumption on the host.  With this change the script tries to create a
dalvik-cache directory on the host if it does not exist.

Change-Id: Ie0c69ff8dcd68b6b04a763f06861edd308406f69