Rename invoke-direct-empty to invoke-object-init
The invoke-direct-empty instruction was introduced to remove the
overhead of calling the empty Object constructor.  We now need it
to do some extra work on behalf of object construction, so it's
appropriate to change the instruction name to match the role it
fills rather than the more general role it was hoped to fill.

No functional changes.

Bug 3342343

Change-Id: I65dd6a2c00c99581c9a19b16fe193b70642c8fbb